+++ lib/concierge.rb
@@ -3,7 +3,6 @@ require 'date'
 require 'csv'
 require 'pry'
 
-require_relative 'lodging'
 require_relative 'room'
 #concierge can ACCESS list of all rooms in hotel = yes
 #concierge can RESERVE reservation for date range = yes
@@ -22,7 +21,7 @@ module Lodging
     attr_accessor :rooms_info
 
     def initialize(room_count)
-      Lodging.create_rooms(room_count) #creating rooms upon initialization
+      Concierge.create_rooms(room_count) #creating rooms upon initialization
 
       rooms = CSV.read('data/all_hotel_rooms.csv', headers: true, header_converters: :symbol).map do |room| #converting string headers to symbols for hash
         room[:reserved_dates] = [] #converting string '[]' to empty array []
@@ -39,9 +38,9 @@ module Lodging
 
       raise ArgumentError if Date.parse(check_out) < Date.parse(check_in) #checkout cannot be earlier than check in
 
-      book_this = Lodging.room_status(@rooms_info, check_in) #returns one available room
+      book_this = Concierge.available_room(@rooms_info, check_in) #returns one available room
 
-      reserved_dates = Lodging.create_date_range(check_in, check_out) #parses range for date, into array
+      reserved_dates = Concierge.create_date_range(check_in, check_out) #parses range for date, into array
 
       update_room = @rooms_info.find do |room|
         room[:room_number] == book_this[:room_number]
@@ -55,10 +54,10 @@ module Lodging
 
       raise ArgumentError if Date.parse(check_out) < Date.parse(check_in) #checkout cannot be earlier than check in
 
-      reserved_dates = Lodging.create_date_range(check_in, check_out) #parses range for date, into array
+      reserved_dates = Concierge.create_date_range(check_in, check_out) #parses range for date, into array
 
       block.times do
-        hold_this = Lodging.room_status(@rooms_info, check_in) #returns one available room
+        hold_this = Concierge.available_room(@rooms_info, check_in) #returns one available room
 
         update_room = @rooms_info.find do |room|
           room[:room_number] == hold_this[:room_number]
@@ -74,7 +73,7 @@ module Lodging
     end
 
     def block_reserve(check_in, check_out)
-      block_dates = Lodging.create_date_range(check_in, check_out) #parses range for date, into array
+      block_dates = Concierge.create_date_range(check_in, check_out) #parses range for date, into array
       held_room = @rooms_info.find do |room|
         room[:status] == 'hold' && room[:reserved_dates] == block_dates
       end #finds room to be reserved
@@ -127,10 +126,41 @@ module Lodging
       cost = check_out_room[:cost].to_f
       total_days = check_out_room[:reserved_dates].length - 1
 
-      return Lodging.total_owed(total_days, cost)
+      return Concierge.total_owed(total_days, cost)
 
     end
 
+    #creates multiple rooms at once, and assigns room number
+    def self.create_rooms(room_count, cost = 200) #creates new rooms, assigns room no.
+      raise ArgumentError if !room_count.is_a? Integer
+      i = 1
+      room_count.times do
+        Room.new(i, cost)
+        i += 1
+      end
+    end
+
+  def self.available_room(input, check_in) #check room_status
+    avail = input.find do |room|
+      room[:reserved_dates].last == Date.parse(check_in) || room[:status] == "available" #returns first instance it finds of available room
+    end
+
+    raise ArgumentError if avail == false
+
+    return avail
+  end
+
+  def self.total_owed(multiplier, price) #date step method to determine day count
+      total = multiplier.to_f * price.to_f
+      return total.round(2)
+  end
+
+  def self.create_date_range(date1, date2)
+    return (Date.parse(date1)..Date.parse(date2)).to_a
+  end
+
+
+  end
 
   end
 end
